+++ packages/feature-state/src/index.ts
@@ -1,5 +1,5 @@
 export * from './create-state';
 export * from './features';
 export * from './is-state-with-features';
-export * from './queue';
+export * from './listener-queue';
 export * from './types';