+++ packages/feature-state/src/types/state.ts
@@ -1,6 +1,6 @@
 import { TWithFeatures, type TFeatureDefinition } from '@blgc/types/features';
 import { type TNestedPath } from '@blgc/utils';
-import type { ListenerQueue } from '../ListenerQueue';
+import type { ListenerQueue } from '../listener-queue';
 
 export type TState<GValue, GFeatures extends TFeatureDefinition[]> = TWithFeatures<
 	{