+++ packages/feature-state/CHANGELOG.md
@@ -1,5 +1,11 @@
 # feature-state
 
+## 0.0.63
+
+### Patch Changes
+
+- updated queue system
+
 ## 0.0.62
 
 ### Patch Changes